requirements.txt
opencv-python~=4.5.5.64
numpy~=1.21.6
pytesseract~=0.3.9
requests~=2.27.1
python-dotenv~=0.20.0
python-telegram-bot~=13.11
